Synchronize audio recordings
Sets up a new way to synchronize recordings. Using Barriers we ensure all
threads are ready, to avoid start up latency from `.play()` we start the
streams. These are running but not writing to file until after the writing flag
is true.

Recordings are stopped via a broadcast channel - each thread is responsible for
tidying it's own streams. With all this in place we reduce app state overhead,
and increase audio synchronization with the discrepancy around 10~20ms.